Prune search space for triples.
We know that a source must be either the result of a function call, the result of a property read, or a parameter (but not an argument to a function call). Similarly, a sanitizer must be a function call, and a sink must be an argument to a function call.

This allows us to drastically restrict the set of nodes we need to check reachability for.